Projeto 1/requirements.txt
py4j==0.10.9.7
pyspark==4.1.1
python-dateutil==2.9.0.post0
six==1.17.0
